edit converter, add list controller, list business logic

+++ src/main/java/kg/academy/maken/service/impl/CardServiceImpl.java
@@ -1,16 +1,14 @@
 package kg.academy.maken.service.impl;
 
-import kg.academy.maken.converter.CardChangeListModelConverter;
+import kg.academy.maken.converter.CardChangeListConverter;
+import kg.academy.maken.converter.CardGetConverter;
 import kg.academy.maken.converter.CardModelConverter;
 import kg.academy.maken.converter.CardPostConverter;
-import kg.academy.maken.entity.Card;
-import kg.academy.maken.entity.CardComment;
-import kg.academy.maken.entity.Comment;
+import kg.academy.maken.entity.*;
 import kg.academy.maken.exception.ApiException;
-import kg.academy.maken.model.CardChangeListModel;
-import kg.academy.maken.model.CardCommentModel;
-import kg.academy.maken.model.CardModel;
-import kg.academy.maken.model.CardPostModel;
+import kg.academy.maken.model.card_model.*;
+import kg.academy.maken.model.commment_model.CommentGetModel;
+import kg.academy.maken.model.commment_model.CommentModel;
 import kg.academy.maken.repository.CardRepository;
 import kg.academy.maken.service.*;
 
@@ -18,6 +16,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -34,15 +34,21 @@ public class CardServiceImpl implements CardService {
     @Autowired
     private CardPostConverter cardPostConverter;
     @Autowired
+    private CardGetConverter cardGetConverter;
+    @Autowired
     private LabelService labelService;
     @Autowired
     private ListService listService;
     @Autowired
-    private CardChangeListModelConverter modelListConverter;
+    private CardChangeListConverter modelListConverter;
     @Autowired
     private CommentService commentService;
     @Autowired
-    private CardCommentService cardCommentService;
+    private CardMemberService cardMemberService;
+    @Autowired
+    private DashboardMemberService memberService;
+    @Autowired
+    private UserService userService;
 
     @Override
     public Card save(Card card) {
@@ -56,13 +62,16 @@ public class CardServiceImpl implements CardService {
 
     @Override
     public Card findById(Long id) {
-        return cardRepository.findById(id).orElse(null);
+        return cardRepository.findById(id)
+                .orElseThrow(() -> new ApiException("Карта не найдена!", HttpStatus.BAD_REQUEST));
     }
 
     @Override
     public Card deleteById(Long id) {
         Card card = findById(id);
-        if (card != null)
+        if (card == null)
+            throw new ApiException("Карта не найдена!", HttpStatus.BAD_REQUEST);
+        else
             cardRepository.deleteById(id);
         return card;
     }
@@ -71,7 +80,7 @@ public class CardServiceImpl implements CardService {
     public List<CardModel> getCardsByList(Long id) {
         List<Card> card = cardRepository.findByList(id).orElse(null);
         if (card == null) {
-            throw new ApiException("Cписок пуст", HttpStatus.NO_CONTENT);
+            throw new ApiException("В листе нет карточек!", HttpStatus.NO_CONTENT);
         }
         return card.stream()
                 .map(cardModelConverter::convertToModel)
@@ -82,7 +91,7 @@ public class CardServiceImpl implements CardService {
     public CardChangeListModel changeList(CardChangeListModel cardChangeListModel) {
         Card card = findById(cardChangeListModel.getListId());
         kg.academy.maken.entity.List list = listService.findById(cardChangeListModel.getListId());
-        if(list== null || card == null)
+        if (list == null || card == null)
             throw new ApiException("Указаны не существующие данные", HttpStatus.BAD_REQUEST);
         else {
             card.setList(list);
@@ -92,14 +101,74 @@ public class CardServiceImpl implements CardService {
     }
 
     @Override
-    public CardCommentModel addComment(CardCommentModel cardCommentModel) {
-        Comment comment = commentService.save(new Comment(cardCommentModel.getComment()));
-        Card card = findById(cardCommentModel.getCardId());
-        CardComment cardComment = new CardComment();
-        cardComment.setComment(comment);
-        cardComment.setComment(comment);
-        cardCommentService.save(cardComment);
-        return null;
+    public CommentModel addComment(CommentModel commentModel) {
+        if (commentModel.getComment() == null)
+            throw new ApiException("Не написан текст!", HttpStatus.BAD_REQUEST);
+        Card card = findById(commentModel.getCardId());
+        User user = userService.findById(commentModel.getCardId());
+        Comment comment = new Comment();
+        comment.setText(commentModel.getComment());
+        comment.setCard(card);
+        comment.setUser(user);
+        commentService.save(comment);
+        return commentModel;
+    }
+
+    @Override
+    public CardMemberModel addMember(CardMemberModel cardMemberModel) {
+        DashboardMember dashboardMember = memberService.findById(cardMemberModel.getDashboardMemberId());
+        Card card = findById(cardMemberModel.getCardId());
+        CardMember cardMember = new CardMember();
+        if (card == null)
+            throw new ApiException("Карточка не найдена", HttpStatus.BAD_REQUEST);
+        cardMember.setDashboardMember(dashboardMember);
+        cardMember.setCard(card);
+        cardMemberService.save(cardMember);
+        return cardMemberModel;
+    }
+
+    @Override
+    public CardMemberModel removeMember(CardMemberModel cardMemberModel) {
+        CardMember cardMember = cardMemberService
+                .findByCardAndMember(cardMemberModel.getCardId(), cardMemberModel.getDashboardMemberId());
+        if (cardMember != null)
+            cardMemberService.deleteById(cardMember.getId());
+        return cardMemberModel;
+    }
+
+    @Override
+    public CardPostModel rejectTask(CardPostModel cardModel) {
+        Card card = findById(cardModel.getID());
+        Long idDashboard = listService.findById(cardModel.getListId()).getDashboard().getId();
+        kg.academy.maken.entity.List list = listService.findByStatusOnDashboard(2L, idDashboard);
+        card.setList(list);
+        cardRepository.save(card);
+        return cardPostConverter.convertToModel(card);
+    }
+
+    @Override
+    public CardRatingModel acceptTask(CardRatingModel cardModel) {
+        Card card = findById(cardModel.getCardId());
+        if (cardModel.getAdminRating() < 0)
+            throw new ApiException("Оценка меньше нуля", HttpStatus.BAD_REQUEST);
+        else {
+            card.setAdminRating(cardModel.getAdminRating());
+            save(card);
+        }
+        return cardModel;
+    }
+
+    @Override
+    public CardGetModel getCard(Long id) {
+        Card card = findById(id);
+        if (card == null)
+            throw new ApiException("Нет такой карточки", HttpStatus.NO_CONTENT);
+        CardGetModel cardGetModel = cardGetConverter.convertToModel(card);
+        List<CommentGetModel> comments = commentService.getByCardId(card.getId());
+        List<CardMemberModel> memberModels = cardMemberService.getMembersByCard(card.getId());
+        cardGetModel.setComments(comments);
+        cardGetModel.setMembers(memberModels);
+        return cardGetModel;
     }
 
     @Override
@@ -111,6 +180,11 @@ public class CardServiceImpl implements CardService {
 
     @Override
     public CardPostModel deleteModelById(Long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        User user = userService.getByLogin(userName);
+        if (!memberService.isAdmin(user, cardRepository.findDashboardByCardId(id)))
+            throw new ApiException("Пользователь не является админом!", HttpStatus.FORBIDDEN);
         return cardPostConverter.convertToModel(deleteById(id));
     }
 
@@ -127,9 +201,9 @@ public class CardServiceImpl implements CardService {
     }
 
     @Override
-    public Page<CardPostModel> getPage(Pageable pageable) {
+    public Page<CardModel> getPage(Pageable pageable) {
         Page<Card> cards = cardRepository.findAll(pageable);
-        return cards.map(cardPostConverter::convertToModel);
+        return cards.map(cardModelConverter::convertToModel);
     }
 
     @Override
@@ -144,7 +218,7 @@ public class CardServiceImpl implements CardService {
             card.setName(model.getName());
         if (model.getDescription() != null)
             card.setDescription(model.getDescription());
-        if(model.getLabelId()!=null)
+        if (model.getLabelId() != null)
             card.setLabel(labelService.findById(model.getLabelId()));
         cardRepository.save(card);
         return cardPostConverter.convertToModel(card);
